requirements.txt
fastapi==0.111.0
SQLAlchemy==2.0.30
uvicorn==0.29.0
pydantic-settings==2.0.3
requests==2.31.0
faker==25.2.0
pytest==8.2.0
fastapi_users==13.0.0